Moved classes into a package
