+++ MyNumber/src/main/java/nikola/games/numberfinder/Calculator.java
@@ -1,0 +1,17 @@
+package nikola.games.numberfinder;
+
+import java.util.ArrayList;
+import jdk.jshell.JShell;
+public class Calculator {
+
+        public static int calculate (String s){
+            int solution=0;
+            JShell js = JShell.builder().build();
+            String sol = (js.eval(s).get(0).value());
+            solution = Integer.parseInt(sol);
+
+
+            return solution;
+        }
+
+}
