+++ MyNumber/src/main/java/nikola/games/numberfinder/Main.java
@@ -1,5 +1,4 @@
-import java.util.ArrayList;
-import java.util.Scanner;
+package nikola.games.numberfinder;
 
 public class Main {
     public static void main(String[] args) {
