+++ MyNumber/src/test/java/nikola/games/numberfinder/CalculatorTest.java
@@ -1,0 +1,18 @@
+package nikola.games.numberfinder;
+
+import nikola.games.numberfinder.Calculator;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class CalculatorTest {
+
+    @Test
+    void calculate() {
+        assertEquals(323, Calculator.calculate("150*2+25-2"));
+        assertEquals(256,Calculator.calculate("50*10/2+8*2-10"));
+
+
+
+    }
+}
